Histogramme für Rot,Grün,Blau und für alle drei Farben gleichzeitig hinzugefügt. Aktuell wird immer noch alles im Arbeitsspeicher abgelegt, was dazu führt, dass dieser nach ein paar Operationen auf großen Bildern volläuft. Das sollte ich auf die Platte auslagern.

+++ picture-editor/contrast.cs
@@ -60,17 +60,17 @@ namespace picture_editor
                     A = origColor.A;// Color.alpha(pixel);
                     // apply filter contrast for every channel R, G, B
                     R = origColor.R;// Color.red(pixel);
-                    R = (int)(((((R / 255.0) - 0.5) * contrast) + 0.5) * 255.0);
+                    R = (int)((((((double)R / 255.0) - 0.5) * contrast) + 0.5) * 255.0);
                     if (R < 0) { R = 0; }
                     else if (R > 255) { R = 255; }
 
                     G = origColor.G;// Color.red(pixel);
-                    G = (int)(((((G / 255.0) - 0.5) * contrast) + 0.5) * 255.0);
+                    G = (int)((((((double)G / 255.0) - 0.5) * contrast) + 0.5) * 255.0);
                     if (G < 0) { G = 0; }
                     else if (G > 255) { G = 255; }
 
                     B = origColor.B;// Color.red(pixel);
-                    B = (int)(((((B / 255.0) - 0.5) * contrast) + 0.5) * 255.0);
+                    B = (int)((((((double)B / 255.0) - 0.5) * contrast) + 0.5) * 255.0);
                     if (B < 0) { B = 0; }
                     else if (B > 255) { B = 255; }
